moved database.js to utils folder
